Started adding a smarter download
The app does (again) check for the images already downloaded to avoid requesting/downloading them again. For the moment the feature is not complete, and will typically not work if the user has downloaded images of a specific rating before. It does also not work for tags with more than 10k images (since images cannot be sort without timeouting the API)